backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
langchain==0.3.7
langchain-community==0.3.7
langchain-ollama==0.2.1
chromadb==0.5.18
sqlalchemy==2.0.35
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
pyjwt==2.10.1
